Moved legacy NetworkInterfaces() implementation details to interfaces.go
